Migrate drop not null constraint requires column data type rule
