+++ src/main/resources/META-INF/services/com.whiteclarkegroup.liquibaselinter.config.rules.ChangeRule
@@ -3,3 +3,4 @@ com.whiteclarkegroup.liquibaselinter.config.rules.core.ModifyDataEnforceWhereImp
 com.whiteclarkegroup.liquibaselinter.config.rules.core.ModifyDataStartsWithWhereImpl
 com.whiteclarkegroup.liquibaselinter.config.rules.core.PrimaryKeyNameRuleImpl
 com.whiteclarkegroup.liquibaselinter.config.rules.core.CreateColumnNullableConstraintRuleImpl
+com.whiteclarkegroup.liquibaselinter.config.rules.core.DropNotNullRequireColumnDataTypeRuleImpl
